requirements/proxy-server.txt
-r ./base.txt

Flask-Assets==0.12
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
psycopg2-binary==2.7.5
python-dotenv==0.9.1
raven[flask]==6.9.0
redis==2.10.6
validators==0.12.2
